+++ plugins/org.projectforge.plugins.datatransfer/src/main/kotlin/org/projectforge/plugins/datatransfer/restPublic/DataTransferPublicServicesRest.kt
@@ -45,9 +45,9 @@ import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.http.ResponseEntity
 import org.springframework.web.bind.annotation.*
 import org.springframework.web.multipart.MultipartFile
-import javax.annotation.PostConstruct
-import javax.servlet.http.HttpServletRequest
-import javax.servlet.http.HttpServletResponse
+import jakarta.annotation.PostConstruct
+import jakarta.servlet.http.HttpServletRequest
+import jakarta.servlet.http.HttpServletResponse
 
 private val log = KotlinLogging.logger {}
 
@@ -84,7 +84,7 @@ class DataTransferPublicServicesRest {
   fun download(
     request: HttpServletRequest,
     @PathVariable("category", required = true) category: String,
-    @PathVariable("id", required = true) id: Int,
+    @PathVariable("id", required = true) id: Long,
     @RequestParam("fileId", required = true) fileId: String,
     @RequestParam("listId") listId: String?,
   )
@@ -123,7 +123,7 @@ class DataTransferPublicServicesRest {
     request: HttpServletRequest,
     response: HttpServletResponse,
     @PathVariable("category", required = true) category: String,
-    @PathVariable("id", required = true) id: Int,
+    @PathVariable("id", required = true) id: Long,
   ): ResponseEntity<*>? {
     check(category == DataTransferPlugin.ID)
     val data = dataTransferPublicSession.checkLogin(request, id) ?: return RestUtils.badRequest("No valid login.")
@@ -159,7 +159,7 @@ class DataTransferPublicServicesRest {
     request: HttpServletRequest,
     response: HttpServletResponse,
     @PathVariable("category", required = true) category: String,
-    @PathVariable("id", required = true) id: Int,
+    @PathVariable("id", required = true) id: Long,
     @RequestParam("fileIds", required = true) fileIds: String,
     @RequestParam("listId") listId: String?
   ) {
@@ -282,7 +282,7 @@ class DataTransferPublicServicesRest {
   fun uploadAttachment(
     request: HttpServletRequest,
     @PathVariable("category", required = true) category: String,
-    @PathVariable("id", required = true) id: Int,
+    @PathVariable("id", required = true) id: Long,
     @PathVariable("listId") listId: String?,
     @RequestParam("file") file: MultipartFile,
   )
@@ -456,7 +456,7 @@ class DataTransferPublicServicesRest {
     request: HttpServletRequest,
     sessionData: DataTransferPublicSession.TransferAreaData,
     category: String,
-    id: Int,
+    id: Long,
     listId: String? = null,
     fileId: String? = null,
     file: Attachment? = null,
