+++ projectforge-business/src/main/kotlin/org/projectforge/business/vacation/model/VacationReplacementFilter.kt
@@ -28,7 +28,7 @@ import org.projectforge.framework.persistence.api.impl.CustomResultFilter
 /**
  * Filters vacation entries by given year.
  */
-class VacationReplacementFilter(val employeeId: Int) : CustomResultFilter<VacationDO> {
+class VacationReplacementFilter(val employeeId: Long) : CustomResultFilter<VacationDO> {
 
   override fun match(list: MutableList<VacationDO>, element: VacationDO): Boolean {
     if (element.replacement?.id == employeeId) {
