+++ projectforge-business/src/main/kotlin/org/projectforge/framework/persistence/api/impl/DBCriteriaContext.kt
@@ -25,13 +25,13 @@ package org.projectforge.framework.persistence.api.impl
 
 import org.projectforge.framework.persistence.api.ExtendedBaseDO
 import org.slf4j.LoggerFactory
-import javax.persistence.criteria.*
+import jakarta.persistence.criteria.*
 
 
 /**
  * Context for building criterias. Holding criteria builder, root path and joinSets.
  */
-internal class DBCriteriaContext<O : ExtendedBaseDO<Int>>(
+internal class DBCriteriaContext<O : ExtendedBaseDO<Long>>(
         val cb: CriteriaBuilder,
         val cr: CriteriaQuery<O>,
         val root: Root<O>,
